chore: shift util functions to utils file
